Create column specificcation

lcols() includes all fields in the input data, skipping the column types as the default.

lcol_lgl(path, .default = zap(), .parser = NULL)

lcol_int(path, .default = zap(), .parser = NULL)

lcol_dbl(path, .default = zap(), .parser = NULL)

lcol_chr(path, .default = zap(), .parser = NULL)

lcol_dat(path, .default = zap(), .parser = NULL)

lcol_dtt(path, .default = zap(), .parser = NULL)

lcol_vec(path, ptype, .default = zap(), .parser = NULL)

lcol_guess(path, .default = NULL)

lcol_skip(path)

lcol_lst(path, .default = zap(), .parser = NULL)

lcol_lst_of(path, .ptype, .default = zap(), .parser = NULL)

lcol_df_lst(path, ..., .default = zap())

lcol_df(path, ..., .default = zap())

Arguments

path
A character vector or list that is converted to an extractor function (similar to the .f argument in purrr::map()).

.default
Value to use if target is empty or absent. If zap() (the default) an error is thrown if target is empty or absent.

.parser
A transformation applied to each element of the list before coercing to .ptype. This is usually needed for lcol_dat() and lcol_dtt().

ptype
The prototype of the vector.

.ptype
The .ptype for vctrs::list_of().

...
Column specification passed on to lcols().