Create a Tibblify Specification

Use spec_df() to specify how to convert a list of objects to a tibble. Use spec_row() resp. spec_object() to specify how to convert an object to a one row tibble resp. a list.

spec_df(..., .names_to = NULL)

spec_object(...)

spec_row(...)

Arguments

...
Column specification created by tib_*() or spec_*().

.names_to
A string giving the name of the column which will contain the names of elements of the object list. If NULL, the default, no name column is created

Examples

spec_df(
  id = tib_int("id"),
  name = tib_chr("name"),
  aliases = tib_chr_vec("aliases")
)
#> spec_df(
#>   id = tib_int("id"),
#>   name = tib_chr("name"),
#>   aliases = tib_chr_vec("aliases")
#> )

# To create multiple columns of the same type use the bang-bang-bang (!!!)
# operator together with `purrr::map()`
spec_df(
  !!!purrr::map(purrr::set_names(c("id", "age")), tib_int),
  !!!purrr::map(purrr::set_names(c("name", "title")), tib_chr)
)
#> spec_df(
#>   id = tib_int("id"),
#>   age = tib_int("age"),
#>   name = tib_chr("name"),
#>   title = tib_chr("title")
#> )

# The `spec_*()` functions can also be nested
spec1 <- spec_object(
  int = tib_int("int"),
  chr = tib_chr("chr")
)
spec2 <- spec_object(
  int2 = tib_int("int2"),
  chr2 = tib_chr("chr2")
)

spec_df(spec1, spec2)
#> spec_df(
#>   int = tib_int("int"),
#>   chr = tib_chr("chr"),
#>   int2 = tib_int("int2"),
#>   chr2 = tib_chr("chr2")
#> )