Guess the tibblify() Specification

Use spec_guess() if you don't know the input type. Use spec_guess_df() if the input is a data frame or an object list. Use spec_guess_objecte() is the input is an object.

spec_guess(
  x,
  ...,
  empty_list_unspecified = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

spec_guess_df(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = TRUE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

spec_guess_object(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = TRUE,
  call = current_call()
)

Arguments

x

A nested list.

...

These dots are for future extensions and must be empty.

empty_list_unspecified

Treat empty lists as unspecified?

inform_unspecified

Inform about fields whose type could not be determined?

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

simplify_list

Should scalar lists be simplified to vectors?

Value

A specification object that can used in tibblify().

Examples

spec_guess(list(x = 1, y = "a"))
#> spec_object(
#>   tib_dbl("x"),
#>   tib_chr("y"),
#> )
spec_guess(list(list(x = 1), list(x = 2)))
#> spec_df(
#>   tib_dbl("x"),
#> )

spec_guess(gh_users)
#> spec_df(
#>   tib_chr("login"),
#>   tib_int("id"),
#>   tib_chr("avatar_url"),
#>   tib_chr("gravatar_id"),
#>   tib_chr("url"),
#>   tib_chr("html_url"),
#>   tib_chr("followers_url"),
#>   tib_chr("following_url"),
#>   tib_chr("gists_url"),
#>   tib_chr("starred_url"),
#>   tib_chr("subscriptions_url"),
#>   tib_chr("organizations_url"),
#>   tib_chr("repos_url"),
#>   tib_chr("events_url"),
#>   tib_chr("received_events_url"),
#>   tib_chr("type"),
#>   tib_lgl("site_admin"),
#>   tib_chr("name"),
#>   tib_chr("company"),
#>   tib_chr("blog"),
#>   tib_chr("location"),
#>   tib_chr("email"),
#>   tib_lgl("hireable"),
#>   tib_chr("bio"),
#>   tib_int("public_repos"),
#>   tib_int("public_gists"),
#>   tib_int("followers"),
#>   tib_int("following"),
#>   tib_chr("created_at"),
#>   tib_chr("updated_at"),
#> )