Create a Field Specification

Use these functions to specify how to convert the fields of an object.

tib_unspecified(key, required = TRUE)

tib_scalar(key, ptype, required = TRUE, default = NULL, transform = NULL)

tib_lgl(key, required = TRUE, default = NULL, transform = NULL)

tib_int(key, required = TRUE, default = NULL, transform = NULL)

tib_dbl(key, required = TRUE, default = NULL, transform = NULL)

tib_chr(key, required = TRUE, default = NULL, transform = NULL)

tib_vector(key, ptype, required = TRUE, default = NULL, transform = NULL)

tib_lgl_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_int_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_dbl_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_chr_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_list(key, required = TRUE, default = NULL, transform = NULL)

tib_row(.key, ..., .required = TRUE)

tib_df(.key, ..., .required = TRUE, .names_to = NULL)

Arguments

key, .key
The path to the field in the object.

required, .required
Throw an error if the field does not exist?

ptype
A prototype of the desired output type of the field.

default
Default value to use if the field does not exist.

transform
A function to apply to the field before casting to ptype.

...
Column specification created by tib_*() or spec_*().

.names_to
A string giving the name of the column which will contain the names of elements of the object list. If NULL, the default, no name column is created

Details

There are basically five different tib_*() functions

tib_scalar(ptype): Cast the field to a length one vector of type ptype.

tib_vector(ptype): Cast the field to an arbitrary length vector of type ptype.

tib_list(): Cast the field to a list.

tib_row(): Cast the field to a named list.

tib_df(): Cast the field to a tibble.

There are some special shortcuts of tib_scalar() resp. tib_vector() for the most common prototypes

logical(): tib_lgl() resp. tib_lgl_vec()

integer(): tib_int() resp. tib_int_vec()

double(): tib_dbl() resp. tib_dbl_vec()

character(): tib_chr() resp. tib_chr_vec()

Examples

tib_int("int")
#> tib_int("int")
tib_int("int", FALSE, default = 0)
#> tib_int(
#>   "int",
#>   required = FALSE,
#>   default = 0
#> )

tib_scalar("date", Sys.Date(), transform = function(x) as.Date(x, format = "%Y-%m-%d"))
#> tib_scalar(
#>   "date",
#>   ptype = vctrs::new_date(),
#>   transform = function (x) 
#>   as.Date(x, format = "%Y-%m-%d")
#> )

tib_df(
  "data",
  .names_to = "id",
  age = tib_int("age"),
  name = tib_chr("name")
)
#> tib_df(
#>   "data",
#>   .names_to = "id",
#>   age = tib_int("age"),
#>   name = tib_chr("name")
#> )