Rectangle a nested list

Rectangle a nested list

Usage

tibblify(x, spec = NULL, names_to = NULL, unspecified = NULL)

Arguments

x

A nested list.

spec

A specification how to convert x. Generated with tspec_row() or tspec_df().

names_to

Deprecated. Use tspec_df(.names_to) instead.

unspecified

A string that describes what happens if the specification contains unspecified fields. Can be one of

"error": Throw an error.

"inform": Inform.

"drop": Do not parse these fields.

"list": Parse an unspecified field into a list.

Value

Either a tibble or a list, depending on the specification

See also

Examples

x <- list(
  list(id = 1, name = "Tyrion Lannister"),
  list(id = 2, name = "Victarion Greyjoy")
)
tibblify(x)
#> # A tibble: 2 × 2
#>      id name             
#>   <dbl> <chr>            
#> 1     1 Tyrion Lannister 
#> 2     2 Victarion Greyjoy

# Provide a specification
spec <- tspec_df(
  id = tib_int("id"),
  name = tib_chr("name")
)
tibblify(x, spec)
#> # A tibble: 2 × 2
#>      id name             
#>   <int> <chr>            
#> 1     1 Tyrion Lannister 
#> 2     2 Victarion Greyjoy

# Provide a specification for a single object
tibblify(x[[1]], tspec_object(spec))
#> $id
#> [1] 1
#> 
#> $name
#> [1] "Tyrion Lannister"
#>